# Conflicts:
#	main/display.c
#	main/gui.c
#	main/input/touchscreen.inc

+++ format.sh
@@ -14,7 +14,7 @@ fi
 (cd main && ${CLANG_FORMAT} -i *.c *.h */*.{c,h,inc})
 pushd libjade
 LIBJADE_SRCS=$(ls *.c *.h | grep -v miniz)
-${CLANG_FORMAT} -i $LIBJADE_SRCS */*.h */*/*.h
+${CLANG_FORMAT} -i $LIBJADE_SRCS selfcheck/*.c */*.h */*/*.h
 popd
 
 ${CLANG_FORMAT} -i tools/bip85_rsa_key_gen/main.c
@@ -24,12 +24,13 @@ if [ -f /.dockerenv ]; then
 fi
 
 if [ -x "$(command -v pycodestyle)" ]; then
-    pycodestyle --max-line-length=100 *.py jadepy/*.py tools/*.py
+    pycodestyle --max-line-length=100 *.py jadepy/*.py tools/*.py tests/*.py tests/*/*.py
 fi
 
 KCONFIG_FILE=main/Kconfig.projbuild
 
 if [ -x ${IDF_PATH}/tools/ci/check_kconfigs.py ]; then
+    rm -f ${KCONFIG_FILE}.new
     ${IDF_PATH}/tools/ci/check_kconfigs.py ${KCONFIG_FILE} || true
-    mv ${KCONFIG_FILE}.new ${KCONFIG_FILE}
+    [ -f ${KCONFIG_FILE}.new ] && mv ${KCONFIG_FILE}.new ${KCONFIG_FILE}
 fi
